requirements/requirements-vqgan-tokens.txt
boto3
pandas
tqdm
numpy
dalle-pytorch
pytorch-lightning==1.7.7
huggingface_hub
pyarrow
